requirements.txt
pymysql
pyyaml